requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.1
fonttools==4.38.0
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.4
psycopg2-binary==2.9.5
py==1.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
seaborn==0.12.1
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
